Bug#17458914 - crash in item_ref::walk, with case, subquery,values() Problem:- It is a regression, cause by the patch for bug(14789787), where INSERT ... ON DUPLICATE KEY UPDATE statement, using VALUES() function with INSERT clause, causes valgrind warnings.
In earlier patch, while resolving VALUES() function, we were checking
every item in UPDATE value list and see whether this VALUES() function 
is in the UPDATE value list.

But there may be chances that some items are not resolved in the UPDATE 
value list, which will result in an assert. This the case with the reported 
bug(17458914), where we are trying to reference the Item_ref in subquery.


So now, instead of traversing the UPDATE item CLAUSE. We will put a new flag, 
which will be set when we are resolving the fields which are on duplicate 
updates.