BUG#17509011: DDL QUERIES INCREASES RPL_SEMI_SYNC_MASTER_YES_TX BY A VALUE OF 2
The problem is that the after_commit hook is being called twice during
the DDL execution: first on the COMMIT stage of the binlog group
commit procedure (MYSQL_BIN_LOG::process_after_commit_stage_queue) and
a second time during the transaction implicit commit
(trans_commit_implicit). And this happens even though the first time
the hook is called, the flag transaction->run_hooks is set to false.

The smallest and quickest fix for this issue is to deploy checks when
we call the after_commit hook, to verify that we have not called the
hook already during this "transaction" lifecycle.