Bug#17463290 - RENAME TABLE AFTER ALTER TABLE MAY NOT BE RECOVERABLE
The combination of ALTER TABLE followed by a RENAME TABLE has a
strange affect if the crash occurs immediately after the RENAME.
There are extra log entries in the 5.6 codepath due to online alter
changes.  But that should not matter since they should only be
replayed once if they are replaying the actual event.  However,
fil_op_log_parse_or_replay() does not check the old name when it
is replayed during recovery.

Another problem is that MLOG_FILE_RENAME entries are replayed when
pages are read in for recovery using asynchronous IO by the the AIO
completion routine if the LSN happens to fit the LSN range of the page.

The combination of these two issues will cause the file to be renamed
back to its original name during recovery

The first fix is to not replay file-based log entries if
fil_op_log_parse_or_replay() is called during IO Completion.

The second fix is to avoid replaying an MLOG_FILE_RENAME log entry
when the current tablespace name is not the same as the old name in
the log entry.

Approved by Marko in RB#3937