Fix for bug#17845161
  ERROR 1296: GOT ERROR 290 'CORRUPT KEY IN TC, UNABLE TO XFRM'; NDB_JOIN_PUSHDOWN

A NULL value in a varchar column was used to construct a lookup key for 
its join-child operation. As a NULL value is never equal to any other value,
such a lookup could have been eliminated already at SPJ. However, in an
attempt to simplify the logic, we instead built a dummy 'empty KeyInfo' for
the NULL lookup. The intention was that such an empty keyInfo would never match
anything and always cause a KEYREF to be returned from LQH.

However, it turned out that such an empty KeyInfo was not sufficiently when
the varchar required usage of key-xfrm to construct an uniform hashable key.
This caused the error message mentioned in bug subject.

This fix use the fact that a NULL-key lookup will never find a matching row.
Instead of sending this useless KEYREQ to a LQH, SPJ will now produce the 
same REF-signals to the API as required by the protocoll.