Bug#17606098: DEADLOCK INVOLVING SRV_MONITOR_THREAD AND LOCK_THD_DATA
Post-push fix: Fix valgrind warnings.

The patch for Bug#17606098 changes the size of the buffer allocated for THD's query string
to match the size of the string. Before it was larger to accomodate data needed for the
query cache. Reducing the size of the buffer triggered valgrind warnings in the lexer
because it sometimes read past the end of the query string. Before this meant reading
into the query cache data, now it means reading past the allocated buffer.

This post-push fix adds a missing eof check to the lexer.