Fix Bug#17033182 FIL_CHECK_FIRST_PAGE() FAILS TO CHECK PAGE SIZE
In fil_check_first_page() do check if the flags are valid and if not
then signal this to the caller by returning != NULL instead of crashing.
In Tablespace::read_lsn_and_check_flags() only read the returned "flags"
and "space" values if the file is reported valid (check_msg == NULL is
returned). If we do not do this, we may hit the assert
"ut_a(!check_tablespace_attributes || space == TRX_SYS_SPACE)" if the
file is corrupted (check_msg != NULL) and thus "space" is bogus.

Reviewed by:    Krunal, Marko, Kevin (rb:4042)