Followup fix for bug#17647637, NDBTICK_CURRENTMICROSECOND() MUST BE MONOTONIC
Previous fix introduced a warning being printed from NdbTick_Init()
if we were unable to find a monotonic timer to be used for the 'ticks'.
OSX is one if the platforms where we have to fallback to the non-monotonic
gettimeofday() and thus prints this warning.

We find from recent PB2 tests, that not only the datanode binaries
ndbd/dbbmtd, but also clients as ndb_mgm need timers, and thus get
this warning printed as part of the startup.

As this is handled as part of the output from these clients, several
tests like 'ndb_mgm -e backup' fails with verification failures.

This fix removes these warning, and instead introduce NdbTick_IsMonotonic().
This function is now used inside Watchdog.cpp to check the timer capabilities
and print a warning from here. As the Watchdog is only part of the ndb's,
this warning will go into the ndb.log file, and not to the terminal.

Patch also adds some extra asserts to NdbTick to
ensure that NdbTick_Init() has been called prior to the other NdbTick
functions, and that the windows part of NdbTick_Init() returned sensible
values. This was added in order to smoke out some Pb2 failures being
observed on Win32/64.