Bug#17600176: This query returns a row in 5.5 but not 5.6 or current 5.7
The problem query contains an IN subquery that is transformed to
a semi-join. The subquery contains an outer join operation.
When run standalone, the outer join in the subquery is preserved as an
outer join and returns one row. However, when the full query is run,
the outer join is converted into an inner join, and the subquery no
longer returns any row. This causes the outer query to return no rows.

The problem is with the IF clause in the WHERE clause of the subquery,
When transforming into semi-join, we rely on Item::fix_after_pullout()
to adjust used_tables and not_null_tables information within the
condition objects, in order to determine e.g. the outer join to inner
join transform. However, there is no specific implementation of
::fix_after_pullout() for the IF clause, so it returns generic
information generated by Item_func::fix_after_pullout(). The fix is to
add this function.

By analysis, this appears to be a problem for BETWEEN predicates and
IN predicates too. A specific implementation of :;fix_after_pullout()
is added for both classes.

In addition, it was detected that not_null_tables information was not
updated correctly for class Item_row. However, I was not able to think
out any failing test for this problem, so no test case was added.