Patch for bug Bug#17947674,
  DEFINTION OF UINT_MAX64 EVALUATES TO A SIGNED VALUE IN VISUAL STUDIO

Also required as an addendum fix to bug#17647637 which introduced usage
of UINT_MAX64.

It seems to be compiler dependent whether the definition of
UINT64_MAX64 actully is *unsigned* as required - In Visual Studio
it actually ends up as a signed value.

This fix makes the UINT_MAX64 more explicit about being an 'unsigned int64'.
Also see how ULONGLONG_MAX is defined in my_global.h.