BUG#17544169 MYSQLBINLOG -V DOES NOT PROPERLY DECODE DECIMAL VALUES IN AN RBR LOG
Analysis: If verbose (-v) option is used in
mysqlbinlog tool it decodes a RBR event into
the query (the same query which was used at
source to generate) and displays immediately
after displaying the row event in the output.
The algorithm used to display the decimal was
wrongly written which were causing problems
similar to the ones mentioned below.
Eg: -0.938582 is decoded as: -938582000.000000000
4294967296.001 as: 000000004.294967296.001000000.000000000 
      
Fix: 'decimal2string' is an existing function
which converts a decimal value into a string
format. Hence the algorithm used was replaced with
this existing 'decimal2string' function to
avoid problems mentioned above.
After this fix, decimal will be printed exactly
the same way how they get printed in 'select' command
output.
Eg: -0.938582 for decimal(10,10) will be displayed as
-0.938582000. 4294967296.001 for decimal(20,10) will
be displayed as 4294967296.0010000000