Fix Bug#17958039 PERFSCHEMA.SIZING_GROWTH FAILING WITH QC ENABLED ON PB2
This is a followup to vasil.dimov@oracle.com-20131203163459-tlkyqdq93jysk9z0.

After adding a per-table rwlock, performance schema needs a minor tweak. It
affects the size of the internal buffer used when instrumenting rwlocks and
will affect the server only if the rwlock instrumentation is used with lot
of tables. The symptoms without this tweak can be to see the rwlock_lost
counter going up (as per Mark Alff).

Reviewed by:	Marc Alff (rb:4184)