Bug #16868967	MYSQL USES SIGNIFICANTLY MORE MEMORY 		FOR ALTER TABLE THAN EXPECTED
Problem:
  During online alter table, an online log buffer would be created for
index that is being rebuilt or created. This online log contains two
buffers - head and tail. Alter table takes lot of memory than
expected even there is no concurrent write for index. But Both head
and tail buffer will be useful when there is a concurrent 
write on the index. If there are many partitions involved, it is
unlikely that all the partitions would be modified concurrently. 
 
Solution:
  Allocate the tail buffer only when first DML statement happens on the
index. Allocate the head buffer only in apply phase and freed at the
end of the apply phase.

	Approved by Marko rb#3850