Fix for bug#17973819  NEED TO ADAPT TO MONOTONIC TIMERS NOT ALWAYS BEING MONOTONIC
Also a followup fix for bug#17647637, NDBTICK_CURRENTMICROSECOND() MUST BE MONOTONIC

This fix handle that Monotonic timers seems to have bugs on several platforms
which might result in the monotonic clock doing small jumps back in time.
This is normally due to imperfect syncing of the clock between multiple 
CPU cores.

Such small backticks are not really harmfull for our scheduler and watchdog
algorithms, so we make the backtick protection less strict in this patch.
However, we will still assert that the backtick is less than 10ms in the
NdbTick_Elapsed() calculation.

We removed redundant checks for backticks in several 'timediff'
methods as such a check already exist in NdbTick_Elapsed() which is 
used in the same place.

We also removed some assert('not backtick') where they are used 
in the same place where NdbTick_Elapsed() was called for the
same reason as above.