Bug#13814468: DYNAMIC RANGE IS SET UP TO BE USED WHEN TYPE               CONVERSION PROHIBITS RANGE ACCESS
Consider a predicate of the form

   WHERE t1.a OP t2.b

where OP is a comparison operator usable by range access. Assume
that the join order is (t1,t2). For table t2, the optimizer may
choose to use access method DYNAMIC RANGE which means that
the range optimizer (and either range access or table/index scan)
is run for each row in t1. However, if t1.a is not comparable to 
t2.b, the range optimizer will not be able to create a range
access plan. In other words, the execution of the range 
optimizer for each row in t1 will be utterly pointless.

The fix is to detect common cases of incompatibility in the 
range optimizer when fields from two tables are compared.