Bug #17991524 THE DEBUG SYNC POINT IB_CORRUPT_PAGE0 IS NOT WORKING AS EXPECTED
Problem:

The debug point ib_corrupt_page0 is not working reliably.  This is because of
the background i/o threads.  If the page 0 of a tablespace is flushed by a user
session (the SET command for variable innodb_buf_flush_list_now), then the
debug point ib_corrupt_page0 will be enabled and will work as expected.  But if
the page 0 of the tablespace is flushed by a background i/o thread, then
ib_corrupt_page0 will not work as expected.  

Solution:

Removing the ib_corrupt_page0 debug point and re-writing the test case using
some other means.  

rb#4216 approved by Jimmy.