WL#6826 - GCS REPLICATION Interfaces for commit validation
This is a post push fix for a possible race condition while the
thread waiting for certification is signalled and the certification
map is updated.
The problem is fixed by moving the mutex_lock to lock the entire 
section of the certification result and the signalling of the wating
thread.