Bug #17209750 OLD FILES NOT BEING REMOVED FROM               PERFORMANCE_SCHEMA.FILE_INSTANCES
Description: "When a file is deleted from the disk, it is
also removed from the file_instances table."  This is not
true on either 5.5.23 or 5.6.11.  When there are three 
relay log files on disk, but the file_instances table 
lists more number of relay logs. Even this case is 
happening for master bin and slave bin logs.

Analysis: When we are deleting the 
slave-relay-bin/master-bin/slave-bin log files we were
using the my_delete() function which is not P_S 
instrumented function. Hence, the files are not removed
from the file_instances table.

Fix: As part of the fix we replaced my_delete() 
function with P_S instrumented mysql_file_delete() function.