Fix for Bug#17857442: MULTITHREADED JOB SCHEDULER SHOULD HANDLE LEAPS IN CLOCK TIME
We have seen customer bug reports, like bug 17475425, where it
looks like delayed signal processing from the 'time_queue' comes
completely out of control: There seems to be two scenarios:

 - Processing of delayes signals stops completely for a long periode,
   or forever. This will cause issues like the disk write speed throttling
   to completely halt any disk activity (Local checkpoints stop)

 - We see a burst of signals which should have been delayed relatively
   to each other, executed withing the same clock second.

Inspecting the mt-scheduler code, we find that this may happen if the
NdbTick clock moves forward or backwards in large leaps. When the time
moved backwards, it will stop the time_queue handling until the
same amount of clock time has passed again. A forward leap will force
the scheduler to simulate that the same amount of time passing by expiring
all time_queue event in the elapsed periode.

Such timer leaps might happen if the platform does not support monotonic
timers, or we didn't implement the usage of these on the specific platform.
Windows used to be such a platform prior to fixing bug 17647637. However,
there are still such platforms, OSX seems to be one of these, and possible
some Linux variants. Furthermore there seems to be several OS/HW/VM bugs
related to monotonic timers not always being monotonic after all. And
there might be time leaps due to CPU starvation such that we are stuck
for a long time - So this has to be expected and handled somehow.

Looking at the single threaded scheduler we find that it actually handle
such leaps - see ThreadConfig::scanTimeQueue(). This logic doesn't seem
to have been ported to the mt-version !! The single threaded handling
is to accept a backtick as the new current time. A forward leap of more
than 1500ms will be consumed by resetting the current time to 'now-1000'ms
and continue from there.

This fix introduce the same handling in the MT-scheduler.

Furthermore, we also found that the mt-scheduler allowed itself to
yield CPU, or wait for receiving signals, even if it has possibly
expired event to handle from the time_queue. 
(see: 'lagging_timers')

We are not sure about whether this has any ill effects - However, 
it doesn't seem like a good idea to let possible expired signals
linger any longer than necessary - also fixed.

Also removed a ndbout_c() which was a leftover from previous debugging.