Bug#17510411: AN INDEPENDENT INSTANCE WITH SEMISYNC ENABLED STILL NEED ACK FROM SLAVE
Problem:
========
When Semi synchronous replication is configured on an
independent server without any slave and
"rpl_semi_sync_master_wait_no_slave" is set to "off" the
master still waits for an ack from the slave.

Analysis:
========
As per the documentation if the value of the variable
"rpl_semi_sync_master_wait_no_slave" is set to OFF, the
master reverts to normal replication if the slave
count drops to zero during the time out period. But in the
existing code the above changes are not implemented.

Fix:
===
When "rpl_semi_sync_master_wait_no_slave" is set to off
check if semi sync is switched on or not if on then check if
any slaves are connected or not. If not connected switch_off
the semi sync and turn it on when user sets the variable to
on.