Fix for bug#18040921 SIGSEV WHEN CONSTRUCTING 'G_THR_REPOSITORY' IF COMPILED WITH -O3
Also intended to fix failing 'Madness' tests in Club2.

'g_thr_repository' was defined and constructed inside mt.cpp as a global data
member.

It is generally adviced against constructing global data members as the order
of construction are somewhat undefined. We change this such that
g_thr_repository is a pointer to a 'struct thr_repository', and
allocate & construct this object as part of the init code instead.