Bug#17959689: MAKE GCC AND CLANG GIVE CONSISTENT COMPILATION WARNINGS
This patch sets the same compiler warning flags when using Clang as
for GCC regardless of debug or release build. With the exception of
-Werror which is still only set for GCC when using MYSQL_MAINTAINER_MODE
(default ON when building debug versions).

This makes it easier to avoid introducing GCC build breaks when using
Clang as compiler (e.g. on OS X).

The patch also moves setting of some warning flags from configure.cmake
to cmake/maintainer.cmake where the rest of the flags are set.

Also note that after this patch, the list of warning flags listed during
configuration is the list of flags actually used for compiling the
server (excluding memcached and cluster).

In addition, the following warning flag changes are done:
* -Wformat-security is added for both GCC/Clang. This is desidered by
  Fedora (https://fedoraproject.org/wiki/Format-Security-FAQ). Compiling
  trunk gives no new warnings with this flag added.

* -Wwrite-strings is only set for C code. It has different meaning for C++
  where it is default on anyway.

* CMake check for -Wdeclaration-after-statement is removed as this flag
  is supported by all current versions of GCC/Clang.

* -Wunused is removed as it included with -Wall and -Wextra.

* -Wno-strict-aliasing is removed as it has no effect when compiling
  with -fno-strict-aliasing. If -fno-strict-aliasing at some point in
  the future is removed, we will very likely want to get strict
  alias warnings reported.

* -Wno-compare-distinct-pointer-types is now set when compiling GUnit
  to silence warnings for code we don't control.

* COMPILE_FLAG_WERROR is removed as it is unused.