Bug#17016214 WASTED WORK IN METHOD LOCK_REC_OTHER_HAS_EXPL_REQ()
Problem: In lock_rec_other_has_expl_req(), we are getting the lock to determine
whether the other trx has the lock request in queue even when heap_no is
PAGE_HEAP_NO_SUPREMUM, which is unnecessary. 

Fixed: No need to check the lock for other trx in the queue, when the heap_no is
PAGE_HEAP_NO_SUPREMUM, as on PAGE_HEAP_NO_SUPREMUM, GAP lock may exist, but we
are not looking for any GAP lock in lock_rec_other_has_expl_req().

Approved by Jimmy (rb#4116).